suppress warnings on stderr
