+++ inkstitch/utils/__init__.py
@@ -1,2 +1,3 @@
 from geometry import *
 from cache import cache
+from io import *
